Move FileHashStore module into HashStore directory and update pytests
