+++ tests/test_filehashstore_interface.py
